section_3/requirements.txt
# AI应用开发核心依赖
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
matplotlib>=3.7.0

# AI模型API
openai>=1.0.0
requests>=2.31.0

# 数据处理
openpyxl>=3.1.0
pillow>=10.0.0

# Web框架
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0

# 数据可视化
dash>=2.14.0
seaborn>=0.12.0
bokeh>=3.2.0

# 部署工具
gunicorn>=21.2.0
docker>=6.1.0

# 环境配置
python-dotenv>=1.0.0

# 开发工具
jupyter>=1.0.0
ipython>=8.14.0

# 测试工具
pytest>=7.4.0
pytest-cov>=4.1.0 